change warehouse.tax    warehouse.ytd       district.tax         district.ytd  from integer to float
git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/regal/trunk/src/jessy@736 334080fa-30e0-4884-a663-39494f8d70c4

+++ benchmark/fr/inria/jessy/benchmark/tpcc/InsertData.java
@@ -27,8 +27,8 @@ public class InsertData extends Transaction {
 				wh = new Warehouse("W_"+i);
 				wh.setW_ID(Integer.toString(i));
 				wh.setW_NAME("Warehouse"+i);
-				wh.setW_TAX((int) rand.nextFloat());
-				wh.setW_YTD((int) rand.nextFloat());
+				wh.setW_TAX(rand.nextFloat());
+				wh.setW_YTD(rand.nextFloat());
 
 				write(wh);
 				
@@ -38,8 +38,8 @@ public class InsertData extends Transaction {
 					dis.setD_ID(Integer.toString(j));
 					dis.setD_W_ID(wh.getW_ID());
 					dis.setD_NAME("District"+j);
-					dis.setD_TAX((int) rand.nextFloat());
-					dis.setD_YTD((int) rand.nextFloat());
+					dis.setD_TAX(rand.nextFloat());
+					dis.setD_YTD(rand.nextFloat());
 					dis.setD_NEXT_O(1);
 
 					write(dis);
